+++ src/form/wmodels.jl
@@ -206,11 +206,9 @@ function constraint_boundary_voltage_angle(pm::_PM.AbstractSOCBFModel, pmd::_PMD
     JuMP.@constraint(pm.model, Wi_to[1,2] == (Wr_to[1,2]*tan(shift_120degs_rad)))
     JuMP.@constraint(pm.model, Wi_to[1,3] == (Wr_to[1,3]*tan(-shift_120degs_rad)))
     JuMP.@constraint(pm.model, Wi_to[2,3] == (Wr_to[2,3]*tan(shift_120degs_rad)))
-
 end
 
 
-# TODO: Find the correct voltage boundary constraint.
 """
     function constraint_boundary_voltage_magnitude(
         pm::_PM.AbstractSOCBFConicModel,
@@ -227,13 +225,12 @@ SOCBFConic-SOCUBFConic boundary bus voltage magnitude (W variables) constraints.
 function constraint_boundary_voltage_magnitude(pm::_PM.AbstractSOCBFConicModel, pmd::_PMD.SOCConicUBFPowerModel, i::Int, f_idx::Tuple{Int,Int,Int}, f_connections::Vector{Int}, t_connections::Vector{Int}; nw::Int = nw_id_default)
     i, f_bus, t_bus = f_idx
 
-    W_fr = _PM.var(pm, nw, :w, f_bus)
-    Wr_to = LinearAlgebra.diag(_PMD.var(pmd, nw, :Wr, t_bus))
-    Wi_to = LinearAlgebra.diag(_PMD.var(pmd, nw, :Wi, t_bus))
+    w_fr = _PM.var(pm, nw, :w, f_bus)
+    wr_to = _PMD.var(pmd, nw, :w, t_bus)
 
-    JuMP.@constraint(pm.model, W_fr == Wr_to[1]+Wi_to[1])
-    JuMP.@constraint(pm.model, Wr_to[2]+Wi_to[2] == Wr_to[1]+Wi_to[1])
-    JuMP.@constraint(pm.model, Wr_to[3]+Wi_to[3] == Wr_to[1]+Wi_to[1])
+    JuMP.@constraint(pm.model, w_fr == wr_to[1])
+    JuMP.@constraint(pm.model, wr_to[1] == wr_to[2])
+    JuMP.@constraint(pm.model, wr_to[1] == wr_to[3])
 end
 
 
@@ -251,10 +248,20 @@ end
 SOCBFConic-SOCUBFConic boundary bus voltage angle constraints.
 """
 function constraint_boundary_voltage_angle(pm::_PM.AbstractSOCBFConicModel, pmd::_PMD.SOCConicUBFPowerModel, i::Int, f_idx::Tuple{Int,Int,Int}, f_connections::Vector{Int}, t_connections::Vector{Int}; nw::Int = nw_id_default)
+    i, f_bus, t_bus = f_idx
+
+    Wr_to = _PMD.var(pmd, nw, :Wr, t_bus)
+    Wi_to = _PMD.var(pmd, nw, :Wi, t_bus)
+
+    shift_120degs_rad = deg2rad(120)
+
+    # These constraints assume angles at the T&D boundary are 0, -120, 120.
+    JuMP.@constraint(pm.model, Wi_to[1,2] == (Wr_to[1,2]*tan(shift_120degs_rad)))
+    JuMP.@constraint(pm.model, Wi_to[1,3] == (Wr_to[1,3]*tan(-shift_120degs_rad)))
+    JuMP.@constraint(pm.model, Wi_to[2,3] == (Wr_to[2,3]*tan(shift_120degs_rad)))
 end
 
 
-# TODO: Find the correct voltage boundary constraint.
 """
     function constraint_boundary_voltage_magnitude(
         pm::_PM.AbstractSOCWRConicModel,
@@ -271,13 +278,12 @@ SOCWRConic-SOCConicUBF boundary bus voltage magnitude (W variables) constraints.
 function constraint_boundary_voltage_magnitude(pm::_PM.AbstractSOCWRConicModel, pmd::_PMD.SOCConicUBFPowerModel, i::Int, f_idx::Tuple{Int,Int,Int}, f_connections::Vector{Int}, t_connections::Vector{Int}; nw::Int = nw_id_default)
     i, f_bus, t_bus = f_idx
 
-    W_fr = _PM.var(pm, nw, :w, f_bus)
-    Wr_to = LinearAlgebra.diag(_PMD.var(pmd, nw, :Wr, t_bus))
-    Wi_to = LinearAlgebra.diag(_PMD.var(pmd, nw, :Wi, t_bus))
+    w_fr = _PM.var(pm, nw, :w, f_bus)
+    wr_to = _PMD.var(pmd, nw, :w, t_bus)
 
-    JuMP.@constraint(pm.model, W_fr == Wr_to[1]+Wi_to[1])
-    JuMP.@constraint(pm.model, Wr_to[2]+Wi_to[2] == Wr_to[1]+Wi_to[1])
-    JuMP.@constraint(pm.model, Wr_to[3]+Wi_to[3] == Wr_to[1]+Wi_to[1])
+    JuMP.@constraint(pm.model, w_fr == wr_to[1])
+    JuMP.@constraint(pm.model, wr_to[1] == wr_to[2])
+    JuMP.@constraint(pm.model, wr_to[1] == wr_to[3])
 end
 
 
@@ -295,11 +301,22 @@ end
 SOCWRConic-SOCConicUBF boundary bus voltage angle constraints.
 """
 function constraint_boundary_voltage_angle(pm::_PM.AbstractSOCWRConicModel, pmd::_PMD.SOCConicUBFPowerModel, i::Int, f_idx::Tuple{Int,Int,Int}, f_connections::Vector{Int}, t_connections::Vector{Int}; nw::Int = nw_id_default)
+
+    i, f_bus, t_bus = f_idx
+
+    Wr_to = _PMD.var(pmd, nw, :Wr, t_bus)
+    Wi_to = _PMD.var(pmd, nw, :Wi, t_bus)
+
+    shift_120degs_rad = deg2rad(120)
+
+    # These constraints assume angles at the T&D boundary are 0, -120, 120.
+    JuMP.@constraint(pm.model, Wi_to[1,2] == (Wr_to[1,2]*tan(shift_120degs_rad)))
+    JuMP.@constraint(pm.model, Wi_to[1,3] == (Wr_to[1,3]*tan(-shift_120degs_rad)))
+    JuMP.@constraint(pm.model, Wi_to[2,3] == (Wr_to[2,3]*tan(shift_120degs_rad)))
 end
 
 
 
-# TODO: Find the correct voltage boundary constraint.
 """
     function constraint_boundary_voltage_magnitude(
         pm::_PM.AbstractSDPWRMModel,
@@ -316,13 +333,12 @@ SDPWRM-SOCConicUBF boundary bus voltage magnitude (W variables) constraints.
 function constraint_boundary_voltage_magnitude(pm::_PM.AbstractSDPWRMModel, pmd::_PMD.SOCConicUBFPowerModel, i::Int, f_idx::Tuple{Int,Int,Int}, f_connections::Vector{Int}, t_connections::Vector{Int}; nw::Int = nw_id_default)
     i, f_bus, t_bus = f_idx
 
-    W_fr = _PM.var(pm, nw, :w, f_bus)
-    Wr_to = LinearAlgebra.diag(_PMD.var(pmd, nw, :Wr, t_bus))
-    Wi_to = LinearAlgebra.diag(_PMD.var(pmd, nw, :Wi, t_bus))
+    w_fr = _PM.var(pm, nw, :w, f_bus)
+    wr_to = _PMD.var(pmd, nw, :w, t_bus)
 
-    JuMP.@constraint(pm.model, W_fr == Wr_to[1]+Wi_to[1])
-    JuMP.@constraint(pm.model, Wr_to[2]+Wi_to[2] == Wr_to[1]+Wi_to[1])
-    JuMP.@constraint(pm.model, Wr_to[3]+Wi_to[3] == Wr_to[1]+Wi_to[1])
+    JuMP.@constraint(pm.model, w_fr == wr_to[1])
+    JuMP.@constraint(pm.model, wr_to[1] == wr_to[2])
+    JuMP.@constraint(pm.model, wr_to[1] == wr_to[3])
 end
 
 
@@ -340,4 +356,15 @@ end
 SDPWRM-SOCConicUBF boundary bus voltage angle constraints.
 """
 function constraint_boundary_voltage_angle(pm::_PM.AbstractSDPWRMModel, pmd::_PMD.SOCConicUBFPowerModel, i::Int, f_idx::Tuple{Int,Int,Int}, f_connections::Vector{Int}, t_connections::Vector{Int}; nw::Int = nw_id_default)
+    i, f_bus, t_bus = f_idx
+
+    Wr_to = _PMD.var(pmd, nw, :Wr, t_bus)
+    Wi_to = _PMD.var(pmd, nw, :Wi, t_bus)
+
+    shift_120degs_rad = deg2rad(120)
+
+    # These constraints assume angles at the T&D boundary are 0, -120, 120.
+    JuMP.@constraint(pm.model, Wi_to[1,2] == (Wr_to[1,2]*tan(shift_120degs_rad)))
+    JuMP.@constraint(pm.model, Wi_to[1,3] == (Wr_to[1,3]*tan(-shift_120degs_rad)))
+    JuMP.@constraint(pm.model, Wi_to[2,3] == (Wr_to[2,3]*tan(shift_120degs_rad)))
 end