+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## staged
 
 - Refactored SOCBF-SOCNLUBF voltage and angle boundary constraints.
+- Refactored SOCBFConic-SOCUBFConic voltage and angle boundary constraints.
+- Refactored SOCWRConic-SOCConicUBF voltage and angle boundary constraints.
+- Refactored SDPWRM-SOCConicUBF voltage and angle boundary constraints.
 
 ## v0.7.5